fix for cached less file with errors
When there is a less compile error when using file caching it makes the
output go blank.
After the error is corrected in the less file the blank file output
keeps being cached.
This was due to response.Cache.SetVaryByCustom("Accept-Encoding") being
called twice because the error output is also written to the response
when using the AspResponseLogger.
When SetVaryByCustom("Accept-Encoding") is called a second time an error
was generated and the file is cached without output and file
dependencies.
It might be a good idea to add a WriteHeaders method to IResponse
interface. This makes it possible for error loggers to append into the
response without affecting the response headers.