Remove try/catch in http handler
The try/catch in the handler hides the detail of the error.  The custom
error settings in the web.config provide sufficient ability to hide
stack traces while allowing developers to perform debugging when needed.
The try/catch also prevents global error handlers from being able to log
the exception.