Allow IE "\9" hack on rules with evaluated values
Bootstrap includes rules such as "background-color: #000 \9; // IE8"
which dotLess would not parse. It was ok with rules such as
"background-color: red \9; // IE8", where it doesn't have to evaluate
the value (it just pulls out the string "red" rather than trying to
parse "#000"). I've added a unit test for the now-fixed evaulated value
version and added one for the already-working non-evaluated value (for
regression testing). All of the existing unit tests still pass. This
relates to Issue 340.